adaptation to weka and some test cases(to be continued)

+++ src/emotionlearner/FeatureExtractorEEG.java
@@ -11,7 +11,7 @@ public class FeatureExtractorEEG extends FeatureExtractor{
 	 * @param data
 	 */
 	@Override
-	public void appendRawData(byte[] data) {
+	public void appendRawData(double[] data) {
 	}
 
 	/**
